Segregate server into multiple subcommands
