+++ osu.Server.DifficultyCalculator/Commands/CommandBase.cs
@@ -1,0 +1,12 @@
+// Copyright (c) 2007-2018 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
+
+using McMaster.Extensions.CommandLineUtils;
+
+namespace osu.Server.DifficultyCalculator.Commands
+{
+    [HelpOption("-?|-h|--help")]
+    public abstract class CommandBase
+    {
+    }
+}
